requirements.txt
streamlit
torch
torchvision
torchaudio
scikit-learn
pandas
numpy
matplotlib
Pillow
gdown
